Update examples.md
I've started providing examples here as it's easier to comprehend something when you've an example to refer to. I've also slightly tweaked the language here and there; hope that's okay. There's more to do but I'm afraid the boys are awake and so I'llhave to stop for now.

PS deliberately used Knockout twice as I thought the contrast was instructive. Happy to include another example of a more standard npm package such as lodash as well