Add reminder on `files -> typings.json`
As mentioned in gitter:

Um….. suddenly I don’t get the purpose of `typings i npm:<pkg>`. If the typings is distributed in the source package, isn’t it safe to assume `package.json/typings` points to the file already thus `tsc` would pick it up automatically?

But if it still make sense, then this PR might worth merging. :smile: